Exam Prepair - Car Service

+++ Exam Preparation - Car Repair Service/skeleton/src/main/java/softuni/exam/models/dto/PartSeedDto.java
@@ -4,14 +4,14 @@ import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.Size;
-import java.math.BigDecimal;
+
 
 public class PartSeedDto {
     @Size(min = 2,max = 19)
     private String partName;
     @Min(10)
     @Max(2000)
-    private BigDecimal price;
+    private double price;
     @Positive
     private int quantity;
 
@@ -23,11 +23,11 @@ public class PartSeedDto {
         this.partName = partName;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
